Clean experiment/common.py
